+++ lib/src/home/controller/counter_controller.dart
@@ -12,22 +12,55 @@ class CounterController extends AppController {
     _model = CounterModel(); // CounterModel(useDouble: true);
 
     /// Provide the 'timer' controller to the interface.
-    wordPairsTimer = WordPairsController();
+    wordPairsTimer = WordPairsTimer();
   }
   static CounterController? _this;
   late final CounterModel _model;
-  late final WordPairsController wordPairsTimer;
+  late final WordPairsTimer wordPairsTimer;
 
   @override
   void initState() {
     super.initState();
+    // Add this controller to the State object's lifecycle.
     wordPairsTimer.addState(state);
+    // initialize the timer
+    wordPairsTimer.initTimer();
   }
 
+  // Merely for demonstration purposes. Erase if not using.
+  /// The framework calls this method whenever it removes this [State] object
+  /// from the tree.
+  @override
+  void deactivate() {
+    super.deactivate();
+  }
+
+  // Merely for demonstration purposes. Erase if not using.
+  /// The framework calls this method when this [State] object will never
+  /// build again.
+  /// Note: THERE IS NO GUARANTEE THIS METHOD WILL RUN in the Framework.
+  @override
+  void dispose() {
+    super.dispose();
+  }
+
+  /// the 'counter' value.
   String get data => _model.data;
 
   /// The 'View' is calling setState()
   void onPressed() => _model.onPressed();
+
+  /// Supply the word pair
+  Widget get wordPair => wordPairsTimer.wordPair;
+
+  /// Access to the timer
+  WordPairsTimer get timer => wordPairsTimer;
+
+  /// Start up the timer
+  void initTimer() => wordPairsTimer.initTimer();
+
+  /// Cancel the timer
+  void cancelTimer() => wordPairsTimer.cancelTimer();
 
   /// The 'Controller' is calling the 'View' to call setState()
 //  void onPressed() => setState(() => _model.onPressed());
@@ -36,54 +69,4 @@ class CounterController extends AppController {
   TemplateController get appController =>
       _appController ??= App.vw!.con as TemplateController;
   TemplateController? _appController;
-
-  /// Start up the timer
-  void initTimer() => wordPairsTimer.initTimer();
-
-  /// Cancel the timer
-  void cancelTimer() => wordPairsTimer.cancelTimer();
-
-  /// Access to the timer
-  WordPairsTimer get timer => wordPairsTimer.timer;
-
-  /// Supply the word pair
-  Widget get wordPair => wordPairsTimer.wordPair;
-
-  /// Supply the app's popupmenu
-  Widget popupMenu({
-    Key? key,
-    List<String>? items,
-    PopupMenuItemBuilder<String>? itemBuilder,
-    String? initialValue,
-    PopupMenuItemSelected<String>? onSelected,
-    PopupMenuCanceled? onCanceled,
-    String? tooltip,
-    double? elevation,
-    EdgeInsetsGeometry? padding,
-    Widget? child,
-    Widget? icon,
-    Offset? offset,
-    bool? enabled,
-    ShapeBorder? shape,
-    Color? color,
-    bool? captureInheritedThemes,
-  }) =>
-      TemplateController().popupMenu(
-        key: key,
-        items: items,
-        itemBuilder: itemBuilder,
-        initialValue: initialValue,
-        onSelected: onSelected,
-        onCanceled: onCanceled,
-        tooltip: tooltip,
-        elevation: elevation,
-        padding: padding,
-        child: child,
-        icon: icon,
-        offset: offset,
-        enabled: enabled,
-        shape: shape,
-        color: color,
-        captureInheritedThemes: captureInheritedThemes,
-      );
 }