+++ lib/src/home/controller/wordpairs_controller.dart
@@ -16,12 +16,6 @@ class WordPairsController extends ControllerMVC {
   final WordPairsTimer timer;
   final WordPairsModel model;
 
-  @override
-  void initState() {
-    super.initState();
-    timer.addState(state);
-  }
-
   /// Start up the timer.
   void initTimer() => timer.initTimer();
 
@@ -39,42 +33,4 @@ class WordPairsController extends ControllerMVC {
   void onTap(int i) => model.onTap(i);
 
   Iterable<Widget> tiles() => model.tiles();
-
-  /// Supply the app's popupmenu
-  Widget popupMenu({
-    Key? key,
-    List<String>? items,
-    PopupMenuItemBuilder<String>? itemBuilder,
-    String? initialValue,
-    PopupMenuItemSelected<String>? onSelected,
-    PopupMenuCanceled? onCanceled,
-    String? tooltip,
-    double? elevation,
-    EdgeInsetsGeometry? padding,
-    Widget? child,
-    Widget? icon,
-    Offset? offset,
-    bool? enabled,
-    ShapeBorder? shape,
-    Color? color,
-    bool? captureInheritedThemes,
-  }) =>
-      TemplateController().popupMenu(
-        key: key,
-        items: items,
-        itemBuilder: itemBuilder,
-        initialValue: initialValue,
-        onSelected: onSelected,
-        onCanceled: onCanceled,
-        tooltip: tooltip,
-        elevation: elevation,
-        padding: padding,
-        child: child,
-        icon: icon,
-        offset: offset,
-        enabled: enabled,
-        shape: shape,
-        color: color,
-        captureInheritedThemes: captureInheritedThemes,
-      );
 }