+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## 2.2.0
+ November 24, 2021
+- Widget popupMenu({
+- appCon.popupMenu()
+- wordPairsTimer = WordPairsTimer();
+- if (state == AppLifecycleState.paused) {
+
 ## 2.1.0+03
  November 18, 2021
 - Introduced .pubignore